composefs: return mkcomposefs stderr as part of error
Writing the error to stderr is not helpful when using the podman API.
The error would only be visable on the server which is bad and likely is
not noticed at all.

Also for the regular podman cli it would be hard for a user to know if
the stderr from mkcompose is related to the returned podman error.

To fix this capture the stderr and append it to the error as string,
this should be fine assuming mkcomposefs doesn't print a ton on stderr
which I assume is not the case.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>