libnetwork: add option to return rootless-netns ips
When using the bridge network mode as rootless we use the rootless netns
logic, for podman this looks like just as using bridge as root. The
issue is however due the extra namespace we block certain address there.
This can be seen best with pasta but actually effects other cases too.
The podman logic tries to use any host ip address for
host.containers.internal but we must make sure to exculde all these
address in the rootless netns as they are not actually the hostns as
thus cause great confusion.

For the --network pasta case I already fixed this by returning the ips on
the pasta.Setup2() call in 83573fa60c0c.
For the bridge mode this more complicated due several layers of function
calls. I decided to implement this as extra function call on the interface
to return the ips as this makes the usage in podman the easiest. And I
also didn't want to break the API as we only have to fix this in podman
not buildah.

It is needed to address #22653 but it needs podman changes as well to
use this new function.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>