Remove nil checks before calling slices.Clone/maps.Clone
Both slices.Clone() and maps.Clone() return `nil` when passed `nil`, so
we don't need to check for a `nil` pointer before calling them.

Imports that used the golang.org/x/exp versions of the maps and slices
packages can now use the versions in the standard library, since we
require Go 1.21.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>