archive: report error from input stream
if there is an error reading from the input stream, prefer to report
it instead of the error from the filter program itself.

We have a test in the buildah CI that expects the "no space left on
device" error that comes from the input stream, to avoid changing the
test, just fix it here.

Reported here: https://github.com/containers/buildah/pull/5585

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>