Use %w for error wrapping
Some code was not using it while it should (this allows a caller to
better inspect the error, if a need arises).

Note in a single case where we have two errors, we only make the
"primary" one unwrappable (by using %w), and explicitly convert the
"secondary" one to a string (which is a way to tell a code reviewer
and a linter that we don't want it to be unwrappable).

Found by errorlint linter.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>