pkg/capabilities: remove init
Replace init and the two global variables with two functions, capList
and capStrList, wrapped in sync.OnceValues.

This way, there is no runtime overhead of merely including this package,
as these functions are executed as needed, not during program
initialization.

While at it:
 - rename getCapName to capName for brewity;
 - use slices.DeleteFunc in capList to filter the slice in place instead
   of making a copy (less garbage to collect);
 - improve AllCapabilities documentation.

Note that currently neither function returns an error; this is a
preparation for the next patch.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>