pkg/netns: ensure makeNetnsDir is race free
There are some rather bad problems when we bind mount over multiple
times, this is a rather small race but can happen. In order to avoid
this take an exclusive lock like ip netns add does because they create
the same bind mount setup. As such we will not race against other podman
process or ip netns add which is a good thing.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>