pkg/netns: do not loop forever
So this is not so simple as one thinks, apparently there are cases where
it is impossible to remove the file but umount() worked fine...
We fixed one issue that ran into this[1] but there seems to be
another[2] problem, unknown cause yet.

Regardless of the real fix for issue[2] add a timeout to not hang/loop
forever. If we were not able to remove the file after 60s give up and
print an error. Leaking these files is not great as the netns references
stay around but it will not prevent containers from running. It will
only start leaking resources.

[1] https://issues.redhat.com/browse/RHEL-59620
[2] https://github.com/containers/podman/issues/24487

Signed-off-by: Paul Holzinger <pholzing@redhat.com>