libnetwork/netavark: do not create config dir in init
Podman creates/initializes the network backend for every command. However
most commands will not need it so we should keep the required actions we
do to a minimum.

In this case the config directory /etc/containers/networks by default as
root may not exists and then we try to create it which can fail, i.e.
when /etc is read only[1].

The code here are a bit more changes then I would have liked but we must
make sure the default in memory network always exists and do not create
the directory there.

[1] https://github.com/containers/common/pull/2265

Signed-off-by: Paul Holzinger <pholzing@redhat.com>