archive: detect symlink target changes
improve detection of changes in symlinks when both the size and the
mtime were changed.

Previously, the detection relied on a difference in either size or
atime/mtime to identify changes in a symlink.  This is not enough, as
the atime/mtime and it size could be left unaltered while only its
target changes.

Fixes: https://github.com/containers/buildah/issues/5861

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>