ignore ENOENT errors when parsing registries.d files
As always listing files in a dir to then read them is racy as the file
might have been removed in the meantime. Thus we must ignore ENOENT
errors when the file is opened.

This is not just a theoretical problem, the reason I am here is because
it caused a flake in the podman CI[1]:
... open /etc/containers/registries.d/podman-test-only-temporary-addition.yaml: no such file or directory

[1] https://api.cirrus-ci.com/v1/artifact/task/6673405799301120/html/int-podman-fedora-40-root-host-boltdb.log.html#t--Podman-push-podman-push-to-local-registry-with-authorization--1

Signed-off-by: Paul Holzinger <pholzing@redhat.com>