rootlessnetns: fix setup error when file already exists
It is possible that the netns file where we bind mount the netns already
exists. This can happen if a previous setup process was killed between
creating the file and mounting to it. Or likely more common as described
in the podman issue if the runroot is not a tmpfs and not deleted after
boot.

Fixes containers/podman#25144

Signed-off-by: Paul Holzinger <pholzing@redhat.com>