pkg/fileutils: add ReflinkOrCopy()
Add a function to copy via reflink if possible or fall back to the
normal io.Copy().
c/image uses this function internally[1] however I like to export this
and use it in podman and I think this package here in c/storage is the
most logical location for it. Note compared to the c/image version I am
calling unix.IoctlFileClone() directly instead of the raw syscall
function.

[1] https://github.com/containers/image/tree/main/internal/reflink

Signed-off-by: Paul Holzinger <pholzing@redhat.com>