pkg/config: fix verify getDefaultEnv test flake
Do not use Default() in unit tests, that caches the config in a global
var so it conflicts with other tests also calling it.

Now only the Reload test should test that. The Default() call got broken
by commit c2dee68766 because now the test only unsets the env after the
last Reload() call which should already use the actual default.

Fixes: c2dee68766 ("*_test.go: use t.TempDir, t.Setenv")

Signed-off-by: Paul Holzinger <pholzing@redhat.com>