Regenerate keys TestKeyFingerprintWithPassphrase and TestKeyFingerprint
... to use 3072-bit RSA keys instead of 1024-bit, and thus hopefully
comply with recent crypto policies without extra setup.

Continue to use RSA for both keys:
- Passphrase-protected ed25519 keys seem to have interoperability issues
  with Sequoia ( https://gitlab.com/sequoia-pgp/sequoia/-/issues/1194 )
- The non-passphrase-protected key needs to be usable in v3 signature packets,
  and those seem not to be defined for ed25519

The v3 signature packets need to be generated using GnuPG < 2.1.0, in
this case it was done 2.0.22 .

It would have been a good idea to script the creation of keys and signatures,
in retrospect, but if the first step is "install a 12-year-old OS",
that would not be _that_ valuable. The next time, if any, we would probably
leave the current RSA key and v3 signatures unchanged, and use a more recent
algorithm for new keys and tests.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>