pkg/config: add systemd wide rootless search locations
Add three new search location when running rootless for the config file:
- /etc/containers/containers.rootless.conf
- /etc/containers/containers.rootless.conf.d
- /etc/containers/containers.rootless.conf.d/<UID>/

The idea is that this can be set by system administrators to define
default config for all rootless users without setting the options for
the rootful podman. Today /etc/containers/containers.conf is already
used as rootless but its options are also read when podman is used as
root which made it not useable for some users.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>