rekor: do not cancel http context
The context must be valid for the entire lifetime of the request
including reading the body, as such it is invalid to return the body
after cancelling the context, depending of if all the response data
arrived already or not this then might fail.

To fix this simply do not cancel the context here, we could move the
cancel() up to the caller function that processes the body. However as
that one already closes the body there should be no need to cancel the
context to end the request.

Fixes: containers/image#2936

Signed-off-by: Paul Holzinger <pholzing@redhat.com>