cirrus: optimize skip/only_if
First a change in .cirrus.yml must run all tests we cannot skip them as
we might chnage what tasks are being executed and how, i.e. an CI image
update must always run all tests.

Second use only_if and not skip, skip will bloat the task overview with
plenty of tasks so it is harder to see what was actully run then.
that is consistent with how we do it in podman.

And drop the vendor task as we do not do vendor anymore.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>