consolidate Makefile validation/tools installs
Create a Makefile in the root of the repo to consolidate how we install
golangci-lint and git-validation, go-md2man. Installing codespell is left
to the user as it is a python program and I didn't want to depend on a
pip install and it wasn't installed anywhere previously either.

Then remove a bunch of targets in the common, image, storage Makefiles
that are done now by the main Makefile.

With that we can also get rid of the custom tools vendor dir in storage.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>