common: golangci-lint fixes for godot and staticcheck
The extra linter which is supposed to run only check fo rnew issues is
broken when the diff is larger than 300 files which can happen.

In general the idea of only flagging new issues doesn't seem that good
compared to just fixing all of them because it is harder to run two
configs locally and most likely never did.

Fix the outstanding issues and remove the extra lint file to fix the CI
issues on PRs which more than 300 files changed.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>