gh action: fix git-validation workflow
In general if we hand a range like 'HEAD~2..HEAD' to git it fails if the
repo only has two commits as HEAD~2 doesn't resolve in such case.
git-validation doesn't seem to fail in such case but it still misses the
first commit.

To fix this we already checked out all PR commits anyway so we can just
validate all commits. But because there is special gh action handling in
git-validation we must disable it as it seem to only check the last
commit based from the GITHUB_ENV var.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>