Update CI image
https://github.com/containers/skopeo/pull/2645 has updated
the expected outputs of Skopeo to match podman-sequoia-0.2.0,
so we need to be running with an updated image that includes that
version, not 0.1.0.

Otherwise, all tests in this repo are failing.

Use /var/tmp for graph driver tests on Debian, because /tmp
is on a tmpfs and we are using an older kernel which does not support
extended attributes on tmpfs.

This adds 20-30 extra seconds to the test runs.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>