storage: remove aufs driver
There are no signs that anyone still uses aufs, no major distro seem to
support it so let's remove it and it hasn't been tested in CI for over
for years[1].

Remove the driver code and config fields for it. There are still a bunch
of references in pkg/archive which I have kept as I am not sure how much
of that still applies.

[1] https://github.com/containers/storage/commit/a54d4e3a4bfbc823e69ef06439ed1cef2658674a

Fixes: #342

Signed-off-by: Paul Holzinger <pholzing@redhat.com>