Test cleartext signature support in gpgme and sequoia backends
Currently, simple signatures are expected to be in the binary GPG
format, and that's what e.g. `podman push --sign-by` produces as well.
But the code for all backends work today with cleartext signatures.

Add a new test to cover this case. But only in the GPGME and Sequoia
backends since the OpenPGP backend does not support it and cannot easily
be supported (see also [1]).

The reason why I'm interested in this is that I'd like to make use
of it for signing Fedora CoreOS container images. The end-goal is to
move to Sigstore signing, but until that's ready, we'd like to use GPG
signing.

We use Robosignatory, the Fedora signing service, which only supports
detached signatures, and while it's theoretically possible to convert
the detached signatures we get back into inline binary signatures,
it's much less cumbersome and error-prone to convert it to cleartext
signatures.

It's worth noting that while Fedora's signing server (Sigul) does
support container image signing, Robosignatory does not surface it yet
(see https://pagure.io/robosignatory/issue/22).

Fixing that wouldn't be too hard I think, but all this is
ideally short-term anyway until we can move to Sigstore
signing + Konflux. There's work in progress on that (see e.g.
https://discussion.fedoraproject.org/t/148999).

The primary goal here is just ensuring that this keeps working until we
move off of it.

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>

[1]: https://github.com/containers/container-libs/pull/307#pullrequestreview-3246302088